+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Changed
+- update intelliJ extension: toggles + panel for include/exclude env vars
+
 ## 3.1.0
 
 ### Added